requirements.txt
git+git://github.com/imankgoyal/etw_pytorch_utils.git@v1.1.1#egg=etw_pytorch_utils
enum34
future
h5py==2.10.0
progressbar2==3.50.0
tensorboardX==2.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.4.0+cu100
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.5.0+cu100
yacs==0.1.6
